requirements.txt
asttokens==2.0.8
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
et-xmlfile==1.1.0
exceptiongroup==1.0.0rc9
executing==1.1.0
h11==0.14.0
icecream==2.1.3
idna==3.4
joblib==1.2.0
lxml==4.9.1
numpy==1.23.3
openpyxl==3.0.10
outcome==1.2.0
pandas==1.5.0
pycparser==2.21
Pygments==2.13.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.1
selenium==4.5.0
six==1.16.0
sklearn==0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
threadpoolctl==3.1.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
webdriver-manager==3.8.3
wsproto==1.2.0
